feat(operations): split files by concept
Split each file into smaller files by concept to allow each code file to be
expanded more easily when more features or options need to be added.
