+++ src/operations/find/sum/mod.rs
@@ -1,7 +1,7 @@
 pub mod pairs;
 
 use clap::Subcommand;
-pub use pairs::Pairs;
+pub use pairs::model::Pairs;
 
 #[derive(Subcommand, Debug)]
 pub enum Sum {
@@ -13,7 +13,7 @@ impl Sum {
     pub fn run(&self) {
         match self {
             Self::Pairs(pairs) => {
-                pairs.run();
+                pairs.run()
             }
         }
     }
