+++ src/operations/find/sum/pairs/solver.rs
